UV_CrewAi/pyproject.toml
[project]
name = "uv-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdul Rehman", email = "121764862+KingAbdulRehman@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.100.1",
    "litellm>=1.59.8",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[project.scripts]
uv-crewai = "uv_crewai:main"
test = "uv_crewai.main:hello"
llmCall = "uv_crewai.call_llm:StartLlmFlow"
StartRoute = "uv_crewai.route_flow:StartRouteFlow"
PlotRoute = "uv_crewai.route_flow:PlotRouteFlow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
